services/rumor_detection/requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7 
torch
Pillow 
patch_ng
lmdb
